requirements.txt
coolsms-python-sdk==2.0.3
